# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Production API
  version: 0.2.0
servers:
  - url: http://127.0.0.1:8710
paths:
...
  /touch/plans/{plan_id}/run:
    post:
      summary: Run a small-batch touch plan against synced contacts
      parameters:
        - in: path
          name: plan_id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Run result }
...